4.21

+++ part04/4-1-bloglist/server/controllers/blogs.js
@@ -28,6 +28,15 @@ blogsRouter.post('/', async (request, response) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const token = request.token;
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' });
+  }
+  const blog = await Blog.findById(request.params.id);
+  if (!blog.user || decodedToken.id !== blog.user.toString()) {
+    return response.status(401).json({ error: 'unauthorized delete' });
+  }
   await Blog.findByIdAndRemove(request.params.id);
   response.status(204).end();
 });
